[graphite] Use SampleCount enum for sample count in TextureInfo
Bug: b/452351267
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/1100357
Commit-Queue: Michael Ludwig <michaelludwig@google.com>

+++ tools/gpu/vk/VkYcbcrSamplerHelper.cpp
@@ -213,7 +213,7 @@ skgpu::VulkanAlloc alloc;
     alloc.fSize = requirements.size;
 
     skgpu::graphite::VulkanTextureInfo imageInfo = {
-            static_cast<uint32_t>(vkImageInfo.samples),
+            vkImageInfo.samples,
             skgpu::Mipmapped::kNo,
             VK_IMAGE_CREATE_PROTECTED_BIT,
             vkImageInfo.format,
